Veeharika's assignment

+++ src/variable/BankBalance.java
@@ -1,5 +1,5 @@
 package variable;
 
 public class BankBalance {
-	double bankBalance;
+	long bankBalance;
 }
